requirements.txt
asgiref==3.2.5
astroid==2.3.3       
beautifulsoup4==4.8.2
cffi==1.14.0
chartjs==1.2
colorama==0.4.1      
cryptography==2.8    
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-audit-log==0.7.0
django-crispy-forms==1.8.1
django-easy-audit==1.2.1
django-import-export==1.2.0
django-model-utils==3.2.0
django-online-users==0.3
django-rest-framework==0.1.0
django-simple-history==2.8.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
html5lib==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.0
openpyxl==3.0.1
pathtools==0.1.2
Pillow==6.2.1
pycparser==2.20
pylint==2.4.4
PyMySQL==0.9.3
PyPDF2==1.26.0
pytz==2019.3
PyYAML==5.1.2
reload==0.9
reportlab==3.5.32
six==1.13.0
soupsieve==2.0
sqlparse==0.3.0
tablib==0.14.0
watchdog==0.10.2
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
dj-database-url==0.5.0
requests==2.23.0
moment